get_paper/requirements.txt
httpx>=0.27.0
feedparser>=6.0.11
pydantic>=2.5.0
python-dateutil>=2.9.0
openreview-py>=1.26.0
openai>=1.52.0
python-dotenv>=1.0.1

